completion {allow,deny} list for filetypes
While it is possible to enable or disable completion for an individual
buffer by setting `b:ollama_enabled`, it is pretty common that some
filetypes just make no sense from the completion standpoint.
Fuzzyfinder windows is my case, but there are probably more.

As a very similar concept, it seems reasonable to assume that some
people would want to enable completion for a limited set of files by
default, and only add new types as they see use for it.